+++ backend/types/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod difficulty;
 pub mod errors;
+pub mod format_strings;
 pub mod lang;
 pub mod pdf;
 pub mod problems;